0000-00-00 is not a valid date. Set it to something more valid.
This will probably need someone caring about it, because if anyone
ever tries to upgrade their MySQL version, I'm not actually sure what
will happen with invalid data like this. It probably will be worth testing.

Probably a good idea to have an upgrade script that updates everything in
the database as part of 3.6?